Refactor database key typing toward identifier-only sublevels
